+++ pyvo/mivot/features/sky_coord_builder.py
@@ -19,17 +19,15 @@ class SkyCoordBuilder:
       set of required parameters (a position).
     - In this implementation, only the mango:EpochPosition class is supported since
       it contains the information required to compute the epoch propagation which is a major use-case
+
+
+    parameters
+    -----------
+    mivot_instance: dict or MivotInstance
+        Python object generated from the MIVOT block as either a Pyhon object or a dict
     '''
 
     def __init__(self, mivot_instance):
-        '''
-        Constructor
-
-        parameters
-        -----------
-        mivot_instance: dict or MivotInstance
-            Python object generated from the MIVOT block as either a Pyhon object or a dict
-        '''
         self._mivot_instance_dict = mivot_instance.to_dict()
         self._map_coord_names = None
 
@@ -108,7 +106,7 @@ class SkyCoordBuilder:
 
         return time_instance
 
-    def _build_time_instance(self, timestamp, representation, besselian=False):
+    def _build_time_instance(self, timestamp, dmtype, besselian=False):
         """
         Build a Time instance matching the input parameters.
         - Returns None if the parameters do not allow any Time setup
@@ -117,9 +115,10 @@ class SkyCoordBuilder:
         parameters
         ----------
         timestamp: string or number
-            The timestamp must comply with the given representation
-        representation: string
-            mango:year, mango:iso, mango:mjd, mango:jd
+            The timestamp must comply with the given dmtype
+        dmtype: string
+            mango:DecimalYear, mango:JulianEpoch, mango:BesselianEpoch,
+            mango:iso, mango:mjd, mango:jd
             (See MANGO primitive types derived from ivoa:timeStamp)
         besselian: boolean (optional)
             Flag telling to use the besselain calendar. We assume it to only be
@@ -128,7 +127,8 @@ class SkyCoordBuilder:
         -------
         Time instance or None
         """
-        if representation in ("mango:year", "yr", "y"):
+        # these types are not MANGO dmtype, however they can encountered in some free-style datasets.
+        if dmtype in ("mango:year", "yr", "y"):
             # it the timestamp is numeric, we infer its format from the besselian flag
             if isinstance(timestamp, numbers.Number):
                 return Time(f"{('B' if besselian else 'J')}{timestamp}",
@@ -155,12 +155,21 @@ class SkyCoordBuilder:
             return None
         # in the following cases, the calendar (B or J) is given by the besselian flag
         # We force to use the  string representation to avoid breaking unit tests.
-        elif representation in ("mango:mjd", "mango:jd", "mango:iso"):
-            astropyformat = representation.split(":")[1]
+        elif dmtype in ("mango:mjd", "mango:jd", "mango:iso"):
+            astropyformat = dmtype.split(":")[1]
             time = Time(f"{timestamp}", format=astropyformat)
             return (Time(time.byear_str) if besselian else time)
+        elif dmtype == "mango:DecimalYear":
+            return Time(float(timestamp), format="decimalyear")
+        elif dmtype == "mango:BesselianYear":
+            if isinstance(timestamp, str) and (timestamp.startswith("B") or timestamp.startswith("J")):
+                timestamp = timestamp[1:]
+            return Time(float(timestamp), format="byear")
+        elif dmtype == "mango:JulianYear":
+            if isinstance(timestamp, str) and (timestamp.startswith("B") or timestamp.startswith("J")):
+                timestamp = timestamp[1:]
+            return Time(float(timestamp), format="jyear")
 
-        print("================= 4")
         return None
 
     def _get_space_frame(self):