+++ pyvo/mivot/tests/test_sky_coord_builder.py
@@ -313,13 +313,32 @@ def test_time_representation():
     scoo = scb.build_sky_coord()
     assert scoo.obstime.jyear_str == "J2025.335"
 
-    mydict = deepcopy(vizier_equin_dict)
-    mydict["obsDate"]["dmtype"] = "mango:year"
-    mydict["obsDate"]["value"] = "B356"
-    mivot_instance = MivotInstance(**mydict)
-    scb = SkyCoordBuilder(mivot_instance)
-    with pytest.raises(MappingError):
-        scb.build_sky_coord()
+    for timestamp in ["2025.0", 2025.0]:
+        mydict = deepcopy(vizier_equin_dict)
+        mydict["obsDate"]["dmtype"] = "mango:DecimalYear"
+        mydict["obsDate"]["value"] = timestamp
+        mivot_instance = MivotInstance(**mydict)
+        scb = SkyCoordBuilder(mivot_instance)
+        scoo = scb.build_sky_coord()
+        assert scoo.obstime.decimalyear == 2025.0
+
+    for timestamp in ["B356", "356", 356.0]:
+        mydict = deepcopy(vizier_equin_dict)
+        mydict["obsDate"]["dmtype"] = "mango:BesselianYear"
+        mydict["obsDate"]["value"] = timestamp
+        mivot_instance = MivotInstance(**mydict)
+        scb = SkyCoordBuilder(mivot_instance)
+        scoo = scb.build_sky_coord()
+        assert pytest.approx(scoo.obstime.decimalyear, 0.1) == 356.0
+
+    for timestamp in ["J1899", "1899", 1899]:
+        mydict = deepcopy(vizier_equin_dict)
+        mydict["obsDate"]["dmtype"] = "mango:JulianYear"
+        mydict["obsDate"]["value"] = timestamp
+        mivot_instance = MivotInstance(**mydict)
+        scb = SkyCoordBuilder(mivot_instance)
+        scoo = scb.build_sky_coord()
+        assert pytest.approx(scoo.obstime.decimalyear, 0.1) == 1899
 
     mydict = deepcopy(vizier_equin_dict)
     mydict["obsDate"]["dmtype"] = "mango:year"