+++ CHANGES.rst
@@ -4,7 +4,9 @@
 Enhancements and Fixes
 ----------------------
 
-- Support the flat representation of the MANGO observation dates (year, mjd, jd, iso) [#726]
+- Support the flat representation of the MANGO observation dates
+  as modeled in the final version of the MANGO data model (DecimalYear,
+  BesselianEpoch, JulianEpoch, mjd, jd, iso) [#726]
 
 - Add DEFAULT_JOB_POLL_TIMEOUT constant [#721]